refactor(skills): 将工具函数与类型抽取到 utils/ 目录
